fix(web): pass buildCuckooBs1 keys in the 2402b16 Uint32Array ABI
The OnionPIR WASM rebuild to onionpir 2402b16 (commit fdca4e24)
surfaced a key-ABI mismatch: 2402b16's buildCuckooBs1 embind wrapper
reads cuckoo keys as a Uint32Array of numHashes*2 (lo32,hi32) pairs
(wasm/hash_utils.cpp), but onionpir_client.ts still passed a
Float64Array of u64 bit-patterns — the pre-2402b16 ABI. The WASM read
garbage keys, built a garbage cuckoo table, and the CHUNK phase failed
with "Entry <id> not in cuckoo table for group <g>".

Replace buildCuckooKeysFloat64 with packCuckooKeysU32 (Uint32Array
lo/hi pairs); use it at both buildCuckooBs1 call sites (chunk cuckoo +
Merkle-sibling cuckoo). Fix the OnionPirModule.buildCuckooBs1 type and
the hand-written onionpir_client.d.ts.

Verified live against wss://pir1.chenweikeng.com: OnionPIR query runs
end-to-end — registration, INDEX, "Level 2 complete: 32 entries
recovered", 2/2 results (1D4HSHPJ... -> 28 UTXOs).

+++ web/public/wasm/onionpir_client.d.ts
@@ -83,13 +83,14 @@ export interface OnionPirModule {
     /**
      * Build a cuckoo hash table for a group of entries.
      * - entries: sorted Uint32Array of entry ids in this group
-     * - keys: Float64Array (treated as uint64s) of 6 hash function keys
+     * - keys: Uint32Array of numHashes*2 elements — each key as a
+     *   consecutive (lo32, hi32) pair
      * - numBins: table size
      * Returns Uint32Array of length numBins; bin == 0xFFFFFFFF means empty.
      */
     buildCuckooBs1(
         entries: Uint32Array,
-        keys: Float64Array,
+        keys: Uint32Array,
         numBins: number,
     ): Uint32Array;
 }
